商家详情评论显示

+++ SVIPApp/src/main/java/com/zkjinshi/svip/activity/shop/ShopDetailActivity.java
@@ -84,7 +84,7 @@ public class ShopDetailActivity extends BaseActivity {
     private RelativeLayout webViewLayout;
     private RatingBar evaluateRb;
     private RelativeLayout evaluateLayout;
-    private int ratingNum;
+    private float ratingNum;
     private String shopId;
     private ShopVo shopVo;
     private int shopStatus;
